change 1 to 0 - test2

+++ src/main/java/com/simon/services/AlgorithmService.java
@@ -41,7 +41,7 @@ public class AlgorithmService {
 
 		array.add(new Integer(1));
 		array.add(new Integer(8));
-		array.add(new Integer(1));
+		array.add(new Integer(0));
 
 		trial.setData(array);
 		trials.add(trial);
